requirements-to-freeze.txt
Django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
pyjwt
cryptography

celery
redis
hiredis

gunicorn
